medwiki_facts_extractor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medwiki-facts-extractor"
version = "0.1.0"
description = "Extract structured facts from medical articles using MiniMax-M2 and DSPy"
requires-python = ">=3.10"
dependencies = [
    "dspy-ai>=2.4.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "aiohttp>=3.9.0",
    "asyncio-timeout>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
